extract flutter watch from flutter analyze
